pyproject.toml
[tool.poetry]
name = "auto_selenium"
version = "0.1.0"
description = ""
authors = ["taimast <81189712+taimast@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.5.0"
webdriver-manager = "^3.8.3"
pydantic = "^1.10.2"
loguru = "^0.6.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
